// react-konva/src/Models/IKeypoints.ts
export interface IKeypoints {
  nose: Array<number>;
  left_eye: Array<number>;
  right_eye: Array<number>;
  left_ear: Array<number>;
  right_ear: Array<number>;
  left_shoulder: Array<number>;
  right_shoulder: Array<number>;
  left_elbow: Array<number>;
  right_elbow: Array<number>;
  left_wrist: Array<number>;
  right_wrist: Array<number>;
  left_hip: Array<number>;
  right_hip: Array<number>;
  left_knee: Array<number>;
  right_knee: Array<number>;
  left_ankle: Array<number>;
  right_ankle: Array<number>;
}